+++ src/App.jsx
@@ -3,7 +3,7 @@ import { Game } from './Game';
 import { BoardWrapper } from './components/BoardWrapper';
 import { parseUrl } from './Parsing';
 import { Footer } from './components/Footer';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 
 const wrapBoardWithReload = ({ reload, board: RawBoard }) => {
 	const Board = (props) => {
@@ -15,20 +15,35 @@ const wrapBoardWithReload = ({ reload, board: RawBoard }) => {
 
 export const App = () => {
 	const { seed, setupData } = parseUrl();
-	const [game, setGame] = useState({...Game(setupData), seed})
+	const [game, setGame] = useState(null)
 
 	const reload = () => {
-		setGame({...Game(setupData), seed});
+		setGame(null);
 	}
 
-	const Client = BgioClient({
+	useEffect(() => {
+		let w;
+		if (game === null) {
+			w = new Worker("src/PuzzleGenWebWorker.js", { type: 'module' });
+			w.postMessage(setupData);
+			w.onmessage = (event) => {
+				// todo, handle errors
+				const { cells, knownCells } = event.data;
+				setGame({...Game({...setupData, cells, knownCells}), seed});
+			};	
+		}
+
+		return () => { if (w) { w.terminate(); }};
+	}, [game, setGame]);
+
+	const Client = game ? BgioClient({
 		game,
 		board: wrapBoardWithReload({ reload, board: BoardWrapper }),
 		numPlayers: 1,
 		debug: {
 			collapseOnLoad: true,
 		}
-	});
+	}) : () => <div>Generating Board...</div>;
 
 	console.log(`Loading game: ${setupData.gamemode === 'c' ? "classic" : "puzzle"} gamemode${seed != null ? ` with a seed of \"${seed}\"` : ""}, ${setupData.count} piece${setupData.count > 1 ? "s" : ""}, ${setupData.size}x${setupData.size} grid, piece${Object.keys(setupData.pieces).length > 1 ? "s" : ""} allowed: ${Object.keys(setupData.pieces).map(x => `${x} (x${setupData.pieces[x]})`).join(', ')}`)
